+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [2.5.0] - 2020-11-11
 ### Added
 - Added sliding strategy
+- 'GuidedSearcher' class which will search for specification with a specific set of 
+  underlying tilings. Includes methods 'from_spec' and 'from_uri' which creates a 
+  'GuidedSearcher' instance.
 - Creates a new `LimitedAssumptionTileScope` that allows you to set a maximum number
     of assumptions allowable on any tiling